==> 03-cleaning.Rmd <==
# Data transformation

```{r, include=FALSE}
source("srcs/helpers.R")
```

In this section, we will discuss how to merge multiple movie data. It can be tricky to merge data from multiple sources. Talk is cheap, click the view source button to see the code if you are interested!

## Input data file

```{r,eval=FALSE}
name.basics = fread(file = 'data/input/name.basics.tsv')
principals = fread(file = 'data/input/title.principals.tsv')
akas = fread(file = 'data/input/title.akas.tsv')
ratings = fread('data/input/title.ratings.tsv')
basics <- fread('data/input/title.basics.tsv')

name.basics %>% head(100) -> name.basics_head
principals %>% head(100) -> principals_head
ratings %>% head(100) -> ratings_head
basics %>% head(100) -> basics_head
akas %>% head(100) -> akas_head

save(name.basics_head, principals_head, ratings_head, basics_head, akas_head,
     file = "data/processing/data_samples.RData")
```


```{r}
load(file = "data/processing/data_samples.RData")
```

- `basics` contains following information for each movies:

```{r}
basics_head %>% 
  get_DT(default_show = 5)
```

  - `akas` contains all alias for each movies:
  
  ```{r}
  akas_head %>% 
    get_DT(default_show = 5)
  ```


- `ratings` contains IMDb rating and votes:

```{r}
ratings_head %>%
  get_DT(default_show = 5)
```

- `principals` contains principal cast/crew for each movies

```{r}
principals_head %>%
  get_DT(default_show = 5)
```

- `Budget` contains rev/cost info

```{r}
budget <- readxl::read_xlsx( 'data/input/budget.xlsx') 
budget %>%
  head(100) %>%
  get_DT(default_show = 5)
```




## Merge & Preprocess Data

### Limit the scope

At this report, we dive in the "movies", and especially those with box office (go the thereter). But the original dataset contains all these 10 types:

```{r,eval=FALSE}
ratings %>% 
  inner_join(basics) %>%
  as_tibble() -> imdb
save(imdb, file = "data/processing/imdb.RData")
```


```{r}
load(file = "data/processing/imdb.RData")
imdb %>%
  #filter(titleType=="tvMovie") %>% 
  group_by(titleType) %>%
  mutate(n = n(),
         top1V = max(numVotes)) %>%
  group_by(titleType) %>%
  arrange(-numVotes) %>%
  slice(1:3) %>%
  summarise(`Number of work` = max(n),
            `Highest Vote` = max(top1V),
            `Top 3 popular voted` = paste0(primaryTitle, collapse = "<br/>")) %>%
  mutate(titleType = ifelse(grepl("Movie", titleType, ignore.case = T), 
                            paste0('<span style="background-color: #277cee44">',
                                   titleType,
                                   '</span>'),
                            titleType)) %>% 
  arrange(-`Number of work`) %>%
get_DT(button = F, default_show = 20)

```

- We only focus on `movie` and `tvMovie` (as some of them have box office, like [High School Musical 3: Senior Year](https://www.the-numbers.com/movies/franchise/High-School-Musical#tab=summary)). And we will use them to map other data.

- Also, we delete movie start at 2022 or later.

### Merge `movie`-wised data

- For basic and rating data, we can use the title id `tconst` to merge directly.

```{r}
imdb %>% 
  filter(titleType %in% c("movie", "tvMovie")) -> imdb_mv
```

- But the for the `budget` data acquired from [numbers](https://www.the-numbers.com/movie/budgets/all/1), we need to map them by movie name and it's alias. Is is a tedious process to locate the IDMB id(`tconst`) value of 6,196  movies in theaters :

  1. Safest way: use movie name and their year to find id
  
        - use `movies name` + `RealeaseYear` in budget to map `primary title` + `year` from `basics` of IDMB. -> ✔️ matches 4495 rows

        - use `movies name` + `RealeaseYear` in budget to map `original title` + `year` from `basics` of IDMB.  -> ✔️ matches 137 rows
    
        - use `movies name`(drop punctuation) + `RealeaseYear` in budget to map `title`(drop punctuation) + `year` from `akas` of IDMB. -> ✔️ matches 410 rows
    
  2. Not the best way: only use its name, and for multiple movies that share a same name, only consider the oen with highest number of votes in IMDB.  -> ✔️ matches 1012 rows
  
  3. Last way: fuzzy matching name + year -  like Star Wars Ep. VII: The Force Awakens with Star Wars: Episode VII -> ✔️ matches 135 rows
  
  4. -> ⚠️ Last 7 movies that can't be matched to IMDB: consider its rev/cost data as missing.
  
  


```{r,eval=FALSE}
akas = fread(file = 'data/input/title.akas.tsv')
budget <- readxl::read_xlsx( 'data/input/budget.xlsx') 
budget %<>%
  mutate(ReleaseDate2 = as.Date(lubridate::parse_date_time(ReleaseDate, "%b %d, %Y")))  %>% 
  mutate(ReleaseDate3 = str_extract(ReleaseDate, "[0-9]+"))  %>% 
  mutate(RealeaseYear = case_when(
    !is.na(ReleaseDate2) ~ year(ReleaseDate2),
    !is.na(ReleaseDate3) ~ as.integer(ReleaseDate3),
    T ~ NA_integer_
  )) %>%
  select(-c(ReleaseDate2, ReleaseDate3)) %>%
  mutate(movie_pure_name = tolower(gsub('[[:space:]]+|[[:punct:]]+', '', Movie)))

akas %>%
   transmute(tconst = titleId, 
             title,
             movie_pure_name = tolower(gsub('[[:space:]]+|[[:punct:]]+', '', title))) %>% 
   inner_join(imdb_mv %>% 
                transmute(tconst,
                          RealeaseYear = as.integer(startYear),
                          numVotes)) %>%
  unique -> akas_mapper


budget_matched = list()
# Safest way: name + year
## use movie primary name + year to map to tconst

budget %>% 
  filter(!is.na(RealeaseYear)) %>%
  inner_join(imdb_mv %>% transmute(tconst, numVotes, Movie = primaryTitle, RealeaseYear = as.integer(startYear))) %>%
  # Resolve conflict
  group_by(Movie, RealeaseYear) %>%
  filter(numVotes == max(numVotes)) %>%
  ungroup -> budget_matched[[1]]

## use movie original name + year to map to tconst
budget %>% 
  anti_join(bind_rows(budget_matched)) %>%
  inner_join(imdb_mv %>% transmute(tconst, numVotes, Movie = originalTitle, RealeaseYear = as.integer(startYear))) %>%
  # Resolve conflict
  group_by(Movie, RealeaseYear) %>%
  filter(numVotes == max(numVotes)) %>%
  ungroup -> budget_matched[[2]]

## use movie akas(drop punct) + year to map to tconst
budget %>%
  anti_join(bind_rows(budget_matched)) %>%
  inner_join(akas_mapper %>% select(tconst, movie_pure_name, RealeaseYear, numVotes)) %>%
  unique %>%
  # Resolve conflict
  group_by(Movie, RealeaseYear) %>%
  filter(numVotes == max(numVotes)) -> budget_matched[[3]]

# Not the best: only use name, and use num Votes to determine which one it refers to.
budget %>%
  anti_join(bind_rows(budget_matched)) %>%
  inner_join(akas_mapper %>% select(tconst, movie_pure_name, numVotes)) %>%
  unique %>% 
  group_by(Movie, RealeaseYear) %>%
  filter(numVotes == max(numVotes)) -> budget_matched[[4]]
  
# Worst: fuzzy matching name + year -  like Star Wars Ep. VII: The Force Awakens with Star Wars: Episode VII
stringdist_join <- function(word){
  tibble(word = word) %>%
  bind_cols(akas_mapper) %>%
    mutate(stringdist = stringdist(word, title, method="osa")) %>%
    filter(stringdist < 0.5 * nchar(word)) %>%
    arrange(-stringdist) -> result
  
  if(nrow(result)==0){
    return(NA)
  }else{
    return(result %>% slice(1) %>% pull(tconst))
  }
}
library(stringdist)
budget %>%
  anti_join(bind_rows(budget_matched)) %>% # takes about 5mins to run!
  rowwise() %>%
  mutate(fuzzy_join_id = stringdist_join(Movie)) -> budget_5


budget_5 %>%
  rename(tconst = fuzzy_join_id) -> budget_matched[[5]]

bind_rows(budget_matched) %>%
  
  distinct(Movie, ReleaseDate, RealeaseYear, tconst, ProductionBudget, DomesticGross, WorldwideGross) %>%
  group_by(tconst) %>%
  filter(RealeaseYear == max(RealeaseYear),
         WorldwideGross == max(WorldwideGross)) %>%
  ungroup -> budget_tconst_mapper
save(budget_tconst_mapper, file="data/processing/budget_tconst_mapper.RData")

```

After a tedious clean, we get this data to map them to imdb database:

```{r}
load(file="data/processing/budget_tconst_mapper.RData")
budget_tconst_mapper %>% 
  head(100) %>%
  get_DT()
```

```{r}
imdb_mv %>%
  left_join(budget_tconst_mapper %>%
              select(-c(Movie, ReleaseDate)) %>%
              mutate_at(c("ProductionBudget", "DomesticGross", "WorldwideGross"), 
                        ~ str_remove_all(., "\\$|,") %>% as.numeric()) %>%
              mutate(Profit = WorldwideGross - ProductionBudget,
                     ROI = WorldwideGross / ProductionBudget) ) %>% as.data.frame() %>% arrange(-numVotes) -> imdb_mv_cleaned
```

Then, we need to understand how the NA is encoded in the dataset. So we calculate 15 most common value among each columns:

```{r}
imdb_mv_cleaned %>%
  mutate_all(~as.character(.)) %>%
  pivot_longer(1:ncol(.)) %>%
  group_by(name, value) %>%
  summarise(nunique = n()) %>%
  group_by(name) %>%
  arrange(-nunique) %>%
  slice(1:15) %>%
  get_DT(default_show = 5)
```

  - Apparently, `\N`, for example, is likely to be one kind of tag for missing value among:

  ```r
  SIGN_of_MISSING <- c('\\N', "", "[]", "	#N/A", "#N/A")
  ```



```{r}
SIGN_of_MISSING <- c('\\N', "", "[]", "	#N/A", "#N/A")
imdb_mv_cleaned %>%
  mutate_all( ~ifelse(. %in% SIGN_of_MISSING, NA, .) ) %>%
  mutate(isAdult = ifelse(isAdult %in% c('0', '1'), isAdult, NA),
         startYear = as.numeric(startYear)) %>%
  filter(startYear <= 2021) %>%
  select(-RealeaseYear) -> imdb_mv_cleaned_2

```





### Merge `person`-wised data

```{r,eval=FALSE}
library(magick)
library(cowplot)

name.basics = fread(file = 'data/input/name.basics.tsv')
principals = fread(file = 'data/input/title.principals.tsv')

principals %>% 
  right_join(imdb_mv_cleaned) %>%
  left_join(name.basics) %>%
  group_by(nconst, primaryName, category) %>%
  summarise(movies_in = n() ,
            total_voted = sum(numVotes, na.rm=T),
            weighted_avg_rating = sum(averageRating * numVotes, na.rm = T) / total_voted) %>%
  ungroup -> imdb_person

imdb_person %>%
  filter(movies_in >= 3,
         category %in% c("director", "actor", "actress")) %>%
  mutate(category = ifelse(category == "director", category, "actor & actress")) %>%
  arrange(-total_voted) %>%
  group_by(category) %>%
  #slice(1:30) %>% 
  ungroup %>%
  split(~category) -> imdb_people

save(imdb_people, file="data/processing/imdb_people.RData")


principals %>% 
  right_join(imdb_mv_cleaned) %>%
  left_join(name.basics) %>%
  filter(category %in% c("director", "actor", "actress"),
         WorldwideGross > 0) -> imdb_peole_movie

save(imdb_peole_movie, file="data/processing/imdb_peole_movie.RData")

load(file="data/processing/imdb_peole_movie.RData")
```

At this part we merge pricinpals to the movies they are in, and calculate the weighted average rating for each actor, actress and director.

```{r}
load(file="data/processing/imdb_people.RData")
bind_rows(imdb_people) %>%
  get_DT(default_show = 5)
```





### Merge `Movie/genres`-wised data

Noted that the `genres` column is not a cleaned column. To preserve this 1-movie-to-multi-genres relation, we 


 
```{r,eval=FALSE}
imdb_mv_cleaned_2 %>% 
  select(tconst, genres, numVotes) %>%
  separate_rows(genres, sep = ",") -> movie_genre_map

movie_genre_map %>%
  left_join(movie_genre_map %>%
              group_by(genres) %>%
              filter(!is.na(genres)) %>%
              summarise(votes = sum(numVotes, na.rm=T))) %>%
  arrange(votes) %>%
  group_by(tconst) %>%
  mutate(genres_num = 1:n()) %>%
  
  ungroup %>%
  mutate(genres_type = case_when(
    genres_num==1 ~ "mainGenres",
    genres_num==2 ~ "secondGenres",
    genres_num==3 ~ "thirdGenres",
    T ~ NA_character_
  )) -> temp

imdb_mv_cleaned_2 %>%
  left_join(temp %>%
              pivot_wider(id_cols = tconst,
                          names_from = genres_type,
                          values_from = genres)) %>%
  select(-genres) -> imdb_pure


imdb_pure %>%
  group_by(mainGenres) %>%
  summarise(n = n()) %>%
  arrange(-n) %>%
  mutate(Genres_type = ifelse(min_rank(-n) <=10, mainGenres, "Others") %>%
           factor %>% fct_relevel("Others", after=Inf)) -> geners_type_map

movie_genre_map %>%
  left_join(imdb_pure %>% select(tconst, averageRating, numVotes)) %>%
  group_by(genres) %>%
  summarise(n = sum(numVotes, na.rm=T)) %>%
  arrange(-n) %>%
  mutate(Genres_type = ifelse(min_rank(-n) <=10, genres, "Others") %>%
           factor %>% fct_relevel("Others", after=Inf)) -> geners_type_map_2

save(movie_genre_map, file = "data/processing/movie_genre_map.RData" )
save(imdb_pure, file = "data/processing/imdb(cleaned).RData")
save(geners_type_map, geners_type_map_2, file = "data/processing/genres_mappers.RData")
```
 
- assign `main genres`, `second genres` and `third genres` based on **reverse order of the popularity of genres** to each movie. For example, if a movie is tagged with Crime, Action, and Adventure. We will label Crime (177m votes) as main genre, Adventure (250m votes) as second, and Action (306m votes) as third.


```{r}
load(file = "data/processing/imdb(cleaned).RData")
imdb_pure %>%
  arrange(-numVotes) %>%
  slice(1:100) %>%
  select(primaryTitle, contains("enre")) %>%
  get_DT(default_show = 5)
```
 
- create a one-to-many mapping table for each movie in order to analysis, e.g, aveage votes for each genres.

```{r}
load(file = "data/processing/movie_genre_map.RData" )
movie_genre_map %>%
  head(100) %>%
  arrange(tconst) %>%
  get_DT(default_show = 5)
```

## Create d3 data

Lastly, we calculate the all-time best movies, the result is stored in `data/processing/d3input.csv`.

```{r,eval=FALSE}
load(file = "data/processing/imdb(cleaned).RData")
load(file = "data/processing/genres_mappers.RData")
budget_tconst_mapper %>%
  mutate_at(c("ProductionBudget", "DomesticGross", "WorldwideGross"), 
            ~ str_remove_all(., "\\$|,") %>% as.numeric()) %>%
  left_join(imdb_pure) %>%
  select(RealeaseYear, Movie, WorldwideGross, genres = mainGenres) -> temp

cum_best_10 = list()
for(year in 1933:2021){
  temp %>%
    filter(RealeaseYear <= year) %>%
    arrange(-WorldwideGross) %>%
    slice(1:10)  %>%
    mutate(RealeaseYear = year) -> cum_best_10[[as.character(year)]]
}

bind_rows(cum_best_10) -> datarows
datarows %<>% 
  left_join(
    datarows %>%
      group_by(genres) %>%
      summarise(n = sum(WorldwideGross, na.rm=T)) %>%
      mutate(genres_type = ifelse(min_rank(-n) <=10, 
                                 genres, 
                                 "Others") %>%
                            factor %>% 
                            fct_relevel("Others", after=Inf)) %>%
      select(-n)) %>%
  mutate(genres = genres_type) %>% select(-genres_type)
  

write.csv(datarows, "data/processing/d3input.csv",row.names=F)
write(c("var movie_data = ["), "data/processing/d3input.txt", append=F)
for(i in 1:(nrow(datarows)-1)){
  datarow = datarows %>% slice(i)
  write(
    paste0('    {year:', pull(datarow, 1),
        ',name:"', pull(datarow, 2),'"',
        ',genre:"', pull(datarow, 4),'"',
        ',box_offices:', pull(datarow, 3),
        '},'),
    "data/processing/d3input.txt",
    append=T
  ) 
}
datarow = datarows %>% slice(nrow(datarows))
write(
  paste0('    {year:', pull(datarow, 1),
      ',name:"', pull(datarow, 2),'"',
      ',genre:"', pull(datarow, 4),'"',
      ',box_offics:', pull(datarow, 3),
      '}'),
  "data/processing/d3input.txt",
  append=T
)
write("];", "data/processing/d3input.txt", append=T)
```
